// src/shared/enums/character-class.ts
export enum CharacterClassEnum {
  GLADIATOR = 'gladiator',
  MOON_LORD = 'moon_lord',
  DESTROYER = 'destroyer',
  BARBARIAN = 'barbarian',
  DARK_AVENGER = 'dark_avenger',
  WIND_WALKER = 'wind_walker',
  TEMPEST = 'tempest',
  SNIPER = 'sniper',
  ARTILLERY = 'artillery',
  GLACIANA = 'glaciana',
  SALEANA = 'saleana',
  OBSCURIA = 'obscuria',
  ILLUMIA = 'illumia',
  SAINT = 'saint',
  INQUISITOR = 'inquisitor',
  GUARDIAN = 'guardian',
  CRUSADER = 'crusader',
  ADEPT = 'adept',
  PHYSICIAN = 'physician',
  SHOOTING_STAR = 'shooting_star',
  GEAR_MASTER = 'gear_master',
  BLADE_DANCER = 'blade_dancer',
  SOUL_EATER = 'soul_eater',
  DARK_SUMMONER = 'dark_summoner',
  SPIRIT_DANCER = 'spirit_dancer',
  RIPPER = 'ripper',
  LIGHT_FURY = 'light_fury',
  RAVEN = 'raven',
  ABYSS_WALKER = 'abyss_walker',
  FLURRY = 'flurry',
  VALKYRIE = 'valkyrie',
}